+++ src/compiler/checker.ts
@@ -10519,7 +10519,7 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
         // function/class/{} initializers are themselves containers, so they won't merge in the same way as other initializers
         const container = getAssignedExpandoInitializer(symbol.valueDeclaration);
         if (container) {
-            const tag = isInJSFile(container) ? getJSDocTypeTag(container) : undefined;
+            const tag = getJSDocTypeTag(container);
             if (tag && tag.typeExpression) {
                 return getTypeFromTypeNode(tag.typeExpression);
             }
@@ -22927,18 +22927,7 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
             return propType;
         }
         if (everyType(type, isTupleType)) {
-            return mapType(type, t => {
-                const tupleType = t as TupleTypeReference;
-                const restType = getRestTypeOfTupleType(tupleType);
-                if (!restType) {
-                    return undefinedType;
-                }
-                if (compilerOptions.noUncheckedIndexedAccess &&
-                    index >= tupleType.target.fixedLength + getEndElementCount(tupleType.target, ElementFlags.Fixed)) {
-                    return getUnionType([restType, undefinedType]);
-                }
-                return restType;
-            });
+            return mapType(type, t => getRestTypeOfTupleType(t as TupleTypeReference) || undefinedType);
         }
         return undefined;
     }
@@ -26284,11 +26273,7 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
                 return declaredType;
             }
             // for (const _ in ref) acts as a nonnull on ref
-            if (
-                isVariableDeclaration(node) &&
-                node.parent.parent.kind === SyntaxKind.ForInStatement &&
-                (isMatchingReference(reference, node.parent.parent.expression) || optionalChainContainsReference(node.parent.parent.expression, reference))
-            ) {
+            if (isVariableDeclaration(node) && node.parent.parent.kind === SyntaxKind.ForInStatement && isMatchingReference(reference, node.parent.parent.expression)) {
                 return getNonNullableTypeIfNeeded(finalizeEvolvingArrayType(getTypeFromFlowType(getTypeAtFlowNode(flow.antecedent))));
             }
             // Assignment doesn't affect reference
@@ -27046,32 +27031,42 @@ export function createTypeChecker(host: TypeCheckerHost): TypeChecker {
                 }
                 return type;
             }
+
+            // Check that right operand is a function type with a prototype property
             const rightType = getTypeOfExpression(expr.right);
             if (!isTypeDerivedFrom(rightType, globalFunctionType)) {
                 return type;
             }
-            const instanceType = mapType(rightType, getInstanceType);
-            // Don't narrow from `any` if the target type is exactly `Object` or `Function`, and narrow
-            // in the false branch only if the target is a non-empty object type.
-            if (isTypeAny(type) && (instanceType === globalObjectType || instanceType === globalFunctionType) ||
-                !assumeTrue && !(instanceType.flags & TypeFlags.Object && !isEmptyAnonymousObjectType(instanceType))) {
+
+            let targetType: Type | undefined;
+            const prototypeProperty = getPropertyOfType(rightType, "prototype" as __String);
+            if (prototypeProperty) {
+                // Target type is type of the prototype property
+                const prototypePropertyType = getTypeOfSymbol(prototypeProperty);
+                if (!isTypeAny(prototypePropertyType)) {
+                    targetType = prototypePropertyType;
+                }
+            }
+
+            // Don't narrow from 'any' if the target type is exactly 'Object' or 'Function'
+            if (isTypeAny(type) && (targetType === globalObjectType || targetType === globalFunctionType)) {
                 return type;
             }
-            return getNarrowedType(type, instanceType, assumeTrue, /*checkDerived*/ true);
-        }
 
-        function getInstanceType(constructorType: Type) {
-            const prototypePropertyType = getTypeOfPropertyOfType(constructorType, "prototype" as __String);
-            if (prototypePropertyType && !isTypeAny(prototypePropertyType)) {
-                return prototypePropertyType;
+            if (!targetType) {
+                const constructSignatures = getSignaturesOfType(rightType, SignatureKind.Construct);
+                targetType = constructSignatures.length ?
+                    getUnionType(map(constructSignatures, signature => getReturnTypeOfSignature(getErasedSignature(signature)))) :
+                    emptyObjectType;
             }
-            const constructSignatures = getSignaturesOfType(constructorType, SignatureKind.Construct);
-            if (constructSignatures.length) {
-                return getUnionType(map(constructSignatures, signature => getReturnTypeOfSignature(getErasedSignature(signature))));
+
+            // We can't narrow a union based off instanceof without negated types see #31576 for more info
+            if (!assumeTrue && rightType.flags & TypeFlags.Union) {
+                const nonConstructorTypeInUnion = find((rightType as UnionType).types, (t) => !isConstructorType(t));
+                if (!nonConstructorTypeInUnion) return type;
             }
-            // We use the empty object type to indicate we don't know the type of objects created by
-            // this constructor function.
-            return emptyObjectType;
+
+            return getNarrowedType(type, targetType, assumeTrue, /*checkDerived*/ true);
         }
 
         function getNarrowedType(type: Type, candidate: Type, assumeTrue: boolean, checkDerived: boolean) {
