+++ src/compiler/diagnosticMessages.json
@@ -5558,22 +5558,6 @@
         "category": "Message",
         "code": 6412
     },
-    "Entering conditional exports.": {
-        "category": "Message",
-        "code": 6413
-    },
-    "Resolved under condition '{0}'.": {
-        "category": "Message",
-        "code": 6414
-    },
-    "Failed to resolve under condition '{0}'.": {
-        "category": "Message",
-        "code": 6415
-    },
-    "Exiting conditional exports.": {
-        "category": "Message",
-        "code": 6416
-    },
 
     "The expected type comes from property '{0}' which is declared here on type '{1}'": {
         "category": "Message",
