+++ tests/baselines/reference/bundlerNodeModules1.trace.json
@@ -4,15 +4,12 @@
     "File '/package.json' does not exist.",
     "Loading module 'dual' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration, JSON.",
     "Found 'package.json' at '/node_modules/dual/package.json'.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './index.js'.",
     "File name '/node_modules/dual/index.js' has a '.js' extension - stripping it.",
     "File '/node_modules/dual/index.ts' does not exist.",
     "File '/node_modules/dual/index.tsx' does not exist.",
     "File '/node_modules/dual/index.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'import'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/dual/index.d.ts', result '/node_modules/dual/index.d.ts'.",
     "======== Module name 'dual' was successfully resolved to '/node_modules/dual/index.d.ts' with Package ID 'dual/index.d.ts@1.0.0'. ========",
     "======== Resolving module 'dual' from '/main.mts'. ========",
