--- tests/baselines/reference/conditionalExportsResolutionFallback(moduleresolution=node16).trace.json
@@ -1,39 +1,0 @@
-[
-    "File '/node_modules/dep/dist/package.json' does not exist.",
-    "Found 'package.json' at '/node_modules/dep/package.json'.",
-    "======== Resolving module 'dep' from '/index.mts'. ========",
-    "Explicitly specified module resolution kind: 'Node16'.",
-    "Resolving in ESM mode with conditions 'node', 'import', 'types'.",
-    "File '/package.json' does not exist.",
-    "Loading module 'dep' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
-    "File '/node_modules/dep/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
-    "Matched 'exports' condition 'import'.",
-    "Using 'exports' subpath '.' with target './dist/index.mjs'.",
-    "File name '/node_modules/dep/dist/index.mjs' has a '.mjs' extension - stripping it.",
-    "File '/node_modules/dep/dist/index.mts' does not exist.",
-    "File '/node_modules/dep/dist/index.d.mts' does not exist.",
-    "Failed to resolve under condition 'import'.",
-    "Saw non-matching condition 'require'.",
-    "Matched 'exports' condition 'types'.",
-    "Using 'exports' subpath '.' with target './dist/index.d.ts'.",
-    "File '/node_modules/dep/dist/index.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'types'.",
-    "Exiting conditional exports.",
-    "Resolving real path for '/node_modules/dep/dist/index.d.ts', result '/node_modules/dep/dist/index.d.ts'.",
-    "======== Module name 'dep' was successfully resolved to '/node_modules/dep/dist/index.d.ts' with Package ID 'dep/dist/index.d.ts@1.0.0'. ========",
-    "File 'package.json' does not exist.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups.",
-    "File 'package.json' does not exist according to earlier cached lookups.",
-    "File '/package.json' does not exist according to earlier cached lookups."
-]
