--- tests/baselines/reference/destructureTupleWithVariableElement.errors.txt
@@ -1,31 +1,0 @@
-tests/cases/compiler/destructureTupleWithVariableElement.ts(9,1): error TS18048: 's1' is possibly 'undefined'.
-tests/cases/compiler/destructureTupleWithVariableElement.ts(10,1): error TS18048: 's2' is possibly 'undefined'.
-tests/cases/compiler/destructureTupleWithVariableElement.ts(18,1): error TS18048: 's5' is possibly 'undefined'.
-
-
-==== tests/cases/compiler/destructureTupleWithVariableElement.ts (3 errors) ====
-    // repro from #52302
-    
-    type NonEmptyStringArray = [string, ...Array<string>]
-    
-    const strings: NonEmptyStringArray = ['one', 'two']
-    const [s0, s1, s2] = strings;
-    
-    s0.toUpperCase()
-    s1.toUpperCase() 
-    ~~
-!!! error TS18048: 's1' is possibly 'undefined'.
-    s2.toUpperCase()
-    ~~
-!!! error TS18048: 's2' is possibly 'undefined'.
-    
-    declare const strings2: [string, ...Array<string>, string]
-    
-    const [s3, s4, s5] = strings2;
-    
-    s3.toUpperCase()
-    s4.toUpperCase() 
-    s5.toUpperCase()
-    ~~
-!!! error TS18048: 's5' is possibly 'undefined'.
-
