--- tests/baselines/reference/destructureTupleWithVariableElement.types
@@ -1,63 +1,0 @@
-=== tests/cases/compiler/destructureTupleWithVariableElement.ts ===
-// repro from #52302
-
-type NonEmptyStringArray = [string, ...Array<string>]
->NonEmptyStringArray : [string, ...string[]]
-
-const strings: NonEmptyStringArray = ['one', 'two']
->strings : [string, ...string[]]
->['one', 'two'] : [string, string]
->'one' : "one"
->'two' : "two"
-
-const [s0, s1, s2] = strings;
->s0 : string
->s1 : string | undefined
->s2 : string | undefined
->strings : [string, ...string[]]
-
-s0.toUpperCase()
->s0.toUpperCase() : string
->s0.toUpperCase : () => string
->s0 : string
->toUpperCase : () => string
-
-s1.toUpperCase() 
->s1.toUpperCase() : string
->s1.toUpperCase : () => string
->s1 : string | undefined
->toUpperCase : () => string
-
-s2.toUpperCase()
->s2.toUpperCase() : string
->s2.toUpperCase : () => string
->s2 : string | undefined
->toUpperCase : () => string
-
-declare const strings2: [string, ...Array<string>, string]
->strings2 : [string, ...string[], string]
-
-const [s3, s4, s5] = strings2;
->s3 : string
->s4 : string | undefined
->s5 : string | undefined
->strings2 : [string, ...string[], string]
-
-s3.toUpperCase()
->s3.toUpperCase() : string
->s3.toUpperCase : () => string
->s3 : string
->toUpperCase : () => string
-
-s4.toUpperCase() 
->s4.toUpperCase() : string
->s4.toUpperCase : () => string
->s4 : string
->toUpperCase : () => string
-
-s5.toUpperCase()
->s5.toUpperCase() : string
->s5.toUpperCase : () => string
->s5 : string | undefined
->toUpperCase : () => string
-
