--- tests/baselines/reference/expandoFunctionContextualTypesJSDocInTs.symbols
@@ -1,11 +1,0 @@
-=== tests/cases/compiler/expandoFunctionContextualTypesJSDocInTs.ts ===
-export function Foo() { }
->Foo : Symbol(Foo, Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 0), Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 25))
-
-// This comment should have no effect; this is a TS file.
-/** @type {never} */
-Foo.bar = () => { };
->Foo.bar : Symbol(Foo.bar, Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 25))
->Foo : Symbol(Foo, Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 0), Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 25))
->bar : Symbol(Foo.bar, Decl(expandoFunctionContextualTypesJSDocInTs.ts, 0, 25))
-
