--- tests/baselines/reference/moduleNoneDynamicImport(target=es2020).js
@@ -1,11 +1,0 @@
-//// [tests/cases/compiler/moduleNoneDynamicImport.ts] ////
-
-//// [a.ts]
-const foo = import("./b");
-
-//// [b.js]
-export default 1;
-
-
-//// [a.js]
-const foo = import("./b");
