--- tests/baselines/reference/moduleNoneDynamicImport(target=es2020).symbols
@@ -1,9 +1,0 @@
-=== /a.ts ===
-const foo = import("./b");
->foo : Symbol(foo, Decl(a.ts, 0, 5))
->"./b" : Symbol("/b", Decl(b.js, 0, 0))
-
-=== /b.js ===
-
-export default 1;
-
