--- tests/baselines/reference/moduleNoneDynamicImport(target=es2020).types
@@ -1,10 +1,0 @@
-=== /a.ts ===
-const foo = import("./b");
->foo : Promise<typeof import("/b")>
->import("./b") : Promise<typeof import("/b")>
->"./b" : "./b"
-
-=== /b.js ===
-
-export default 1;
-
