+++ tests/baselines/reference/nodeModulesExportsBlocksTypesVersions(module=nodenext).trace.json
@@ -39,12 +39,9 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/yep' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'types'.",
     "Using 'exports' subpath './yep' with target './types/foo.d.ts'.",
     "File '/node_modules/exports-and-types-versions/types/foo.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'types'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/exports-and-types-versions/types/foo.d.ts', result '/node_modules/exports-and-types-versions/types/foo.d.ts'.",
     "======== Module name 'exports-and-types-versions/yep' was successfully resolved to '/node_modules/exports-and-types-versions/types/foo.d.ts' with Package ID 'exports-and-types-versions/types/foo.d.ts@1.0.0'. ========",
     "======== Resolving module 'exports-and-types-versions/versioned-yep' from '/main.cts'. ========",
@@ -53,12 +50,9 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/versioned-yep' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'types@>=4'.",
     "Using 'exports' subpath './versioned-yep' with target './types/foo.d.ts'.",
     "File '/node_modules/exports-and-types-versions/types/foo.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'types@>=4'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/exports-and-types-versions/types/foo.d.ts', result '/node_modules/exports-and-types-versions/types/foo.d.ts'.",
     "======== Module name 'exports-and-types-versions/versioned-yep' was successfully resolved to '/node_modules/exports-and-types-versions/types/foo.d.ts' with Package ID 'exports-and-types-versions/types/foo.d.ts@1.0.0'. ========",
     "======== Resolving module 'exports-and-types-versions/versioned-nah' from '/main.cts'. ========",
@@ -67,15 +61,11 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/versioned-nah' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Saw non-matching condition 'types@<4'.",
-    "Exiting conditional exports.",
     "Export specifier './versioned-nah' does not exist in package.json scope at path '/node_modules/exports-and-types-versions'.",
     "Directory '/node_modules/@types' does not exist, skipping all lookups in it.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Saw non-matching condition 'types@<4'.",
-    "Exiting conditional exports.",
     "Export specifier './versioned-nah' does not exist in package.json scope at path '/node_modules/exports-and-types-versions'.",
     "======== Module name 'exports-and-types-versions/versioned-nah' was not resolved. ========",
     "======== Resolving module 'just-types-versions/foo' from '/main.cts'. ========",
@@ -136,12 +126,9 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/yep' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'types'.",
     "Using 'exports' subpath './yep' with target './types/foo.d.ts'.",
     "File '/node_modules/exports-and-types-versions/types/foo.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'types'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/exports-and-types-versions/types/foo.d.ts', result '/node_modules/exports-and-types-versions/types/foo.d.ts'.",
     "======== Module name 'exports-and-types-versions/yep' was successfully resolved to '/node_modules/exports-and-types-versions/types/foo.d.ts' with Package ID 'exports-and-types-versions/types/foo.d.ts@1.0.0'. ========",
     "======== Resolving module 'exports-and-types-versions/versioned-yep' from '/main.mts'. ========",
@@ -150,12 +137,9 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/versioned-yep' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'types@>=4'.",
     "Using 'exports' subpath './versioned-yep' with target './types/foo.d.ts'.",
     "File '/node_modules/exports-and-types-versions/types/foo.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'types@>=4'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/exports-and-types-versions/types/foo.d.ts', result '/node_modules/exports-and-types-versions/types/foo.d.ts'.",
     "======== Module name 'exports-and-types-versions/versioned-yep' was successfully resolved to '/node_modules/exports-and-types-versions/types/foo.d.ts' with Package ID 'exports-and-types-versions/types/foo.d.ts@1.0.0'. ========",
     "======== Resolving module 'exports-and-types-versions/versioned-nah' from '/main.mts'. ========",
@@ -164,15 +148,11 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'exports-and-types-versions/versioned-nah' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Saw non-matching condition 'types@<4'.",
-    "Exiting conditional exports.",
     "Export specifier './versioned-nah' does not exist in package.json scope at path '/node_modules/exports-and-types-versions'.",
     "Directory '/node_modules/@types' does not exist, skipping all lookups in it.",
     "File '/node_modules/exports-and-types-versions/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Saw non-matching condition 'types@<4'.",
-    "Exiting conditional exports.",
     "Export specifier './versioned-nah' does not exist in package.json scope at path '/node_modules/exports-and-types-versions'.",
     "======== Module name 'exports-and-types-versions/versioned-nah' was not resolved. ========",
     "======== Resolving module 'just-types-versions/foo' from '/main.mts'. ========",
