+++ tests/baselines/reference/resolvesWithoutExportsDiagnostic1(moduleresolution=bundler).trace.json
@@ -4,24 +4,18 @@
     "File '/package.json' does not exist.",
     "Loading module 'foo' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration, JSON.",
     "Found 'package.json' at '/node_modules/foo/package.json'.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './index.mjs'.",
     "File name '/node_modules/foo/index.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/foo/index.mts' does not exist.",
     "File '/node_modules/foo/index.d.mts' does not exist.",
-    "Failed to resolve under condition 'import'.",
     "Saw non-matching condition 'require'.",
-    "Exiting conditional exports.",
     "File '/node_modules/@types/foo.d.ts' does not exist.",
     "File '/node_modules/foo/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './index.mjs'.",
     "File name '/node_modules/foo/index.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/foo/index.mjs' exists - use it as a name resolution result.",
-    "Resolved under condition 'import'.",
-    "Exiting conditional exports.",
     "Resolution of non-relative name failed; trying with modern Node resolution features disabled to see if npm library needs configuration update.",
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'foo' from 'node_modules' folder, target file types: TypeScript, Declaration.",
@@ -40,27 +34,19 @@
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'bar' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration, JSON.",
     "Found 'package.json' at '/node_modules/bar/package.json'.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './index.mjs'.",
     "File name '/node_modules/bar/index.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/bar/index.mts' does not exist.",
     "File '/node_modules/bar/index.d.mts' does not exist.",
-    "Failed to resolve under condition 'import'.",
     "Saw non-matching condition 'require'.",
-    "Exiting conditional exports.",
     "Found 'package.json' at '/node_modules/@types/bar/package.json'.",
-    "Entering conditional exports.",
     "Saw non-matching condition 'require'.",
-    "Exiting conditional exports.",
     "File '/node_modules/bar/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './index.mjs'.",
     "File name '/node_modules/bar/index.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/bar/index.mjs' exists - use it as a name resolution result.",
-    "Resolved under condition 'import'.",
-    "Exiting conditional exports.",
     "Resolution of non-relative name failed; trying with modern Node resolution features disabled to see if npm library needs configuration update.",
     "File '/package.json' does not exist according to earlier cached lookups.",
     "Loading module 'bar' from 'node_modules' folder, target file types: TypeScript, Declaration.",
