+++ tests/baselines/reference/selfNameModuleAugmentation.trace.json
@@ -3,13 +3,11 @@
     "Explicitly specified module resolution kind: 'Bundler'.",
     "File '/node_modules/acorn-walk/dist/package.json' does not exist.",
     "Found 'package.json' at '/node_modules/acorn-walk/package.json'.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './dist/walk.mjs'.",
     "File name '/node_modules/acorn-walk/dist/walk.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/acorn-walk/dist/walk.mts' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.d.mts' does not exist.",
-    "Failed to resolve under condition 'import'.",
     "Saw non-matching condition 'require'.",
     "Matched 'exports' condition 'default'.",
     "Using 'exports' subpath '.' with target './dist/walk.js'.",
@@ -17,8 +15,6 @@
     "File '/node_modules/acorn-walk/dist/walk.ts' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.tsx' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'default'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/acorn-walk/dist/walk.d.ts', result '/node_modules/acorn-walk/dist/walk.d.ts'.",
     "======== Module name 'acorn-walk' was successfully resolved to '/node_modules/acorn-walk/dist/walk.d.ts' with Package ID 'acorn-walk/dist/walk.d.ts@8.2.0'. ========",
     "======== Resolving module 'acorn-walk' from '/index.ts'. ========",
@@ -26,13 +22,11 @@
     "File '/package.json' does not exist.",
     "Loading module 'acorn-walk' from 'node_modules' folder, target file types: TypeScript, JavaScript, Declaration, JSON.",
     "File '/node_modules/acorn-walk/package.json' exists according to earlier cached lookups.",
-    "Entering conditional exports.",
     "Matched 'exports' condition 'import'.",
     "Using 'exports' subpath '.' with target './dist/walk.mjs'.",
     "File name '/node_modules/acorn-walk/dist/walk.mjs' has a '.mjs' extension - stripping it.",
     "File '/node_modules/acorn-walk/dist/walk.mts' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.d.mts' does not exist.",
-    "Failed to resolve under condition 'import'.",
     "Saw non-matching condition 'require'.",
     "Matched 'exports' condition 'default'.",
     "Using 'exports' subpath '.' with target './dist/walk.js'.",
@@ -40,8 +34,6 @@
     "File '/node_modules/acorn-walk/dist/walk.ts' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.tsx' does not exist.",
     "File '/node_modules/acorn-walk/dist/walk.d.ts' exists - use it as a name resolution result.",
-    "Resolved under condition 'default'.",
-    "Exiting conditional exports.",
     "Resolving real path for '/node_modules/acorn-walk/dist/walk.d.ts', result '/node_modules/acorn-walk/dist/walk.d.ts'.",
     "======== Module name 'acorn-walk' was successfully resolved to '/node_modules/acorn-walk/dist/walk.d.ts' with Package ID 'acorn-walk/dist/walk.d.ts@8.2.0'. ========"
 ]
