+++ tests/baselines/reference/tscWatch/moduleResolution/diagnostics-from-cache.js
@@ -39,12 +39,9 @@ Found 'package.json' at '/user/username/projects/myproject/package.json'.
 Explicitly specified module resolution kind: 'NodeNext'.
 Resolving in ESM mode with conditions 'node', 'import', 'types'.
 File '/user/username/projects/myproject/package.json' exists according to earlier cached lookups.
-Entering conditional exports.
 Matched 'exports' condition 'default'.
 Using 'exports' subpath '.' with target './dist/index.js'.
 File '/user/username/projects/myproject/index.ts' exists - use it as a name resolution result.
-Resolved under condition 'default'.
-Exiting conditional exports.
 Resolving real path for '/user/username/projects/myproject/index.ts', result '/user/username/projects/myproject/index.ts'.
 ======== Module name '@this/package' was successfully resolved to '/user/username/projects/myproject/index.ts'. ========
 File '/user/username/projects/myproject/package.json' exists according to earlier cached lookups.
