--- tests/cases/compiler/expandoFunctionContextualTypesJSDocInTs.ts
@@ -1,5 +1,0 @@
-export function Foo() { }
-
-// This comment should have no effect; this is a TS file.
-/** @type {never} */
-Foo.bar = () => { };
